% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids2names.R
\name{ids2names}
\alias{ids2names}
\title{Given a set of reference IDs, corresponding reference names, and a vector of query IDs, get a vector of common names for the input IDs (or the original input IDs for those IDs not in the reference).}
\usage{
ids2names(ref_ids, ref_names, input_ids)
}
\arguments{
\item{ref_ids}{A vector of ids}

\item{ref_names}{A vector, the same length as ref_ids}

\item{input_ids}{A vector containing ids that will be matched against the reference ids.}
}
\value{
A string vector the same length and order as input_ids. Each element is the reference name if the given input id was found among the reference ids, otherwise it will be the original input id.
}
\description{
Given a set of reference IDs, corresponding reference names, and a vector of query IDs, get a vector of common names for the input IDs (or the original input IDs for those IDs not in the reference).
}
\examples{
gene_names_or_ids <- ids2names(
                           ref_ids=c("id1","id2"),
                           ref_names=c("a","b"),
                           input_ids=c("id1","id3","id2","id1")
                         )
print(gene_names_or_ids) # outputs [1] "a"   "id3" "b"   "a"
}
